Added framework for database service.
The function names have been declared, and their functionality is stated in
comments of the databaseInitialization file. This service will provide an
interface to the sqlite3 database. We still need to obviously add the
functionality, (I also might of missed some arguments to some functions,
and global config variables) and I will design the sqlite3 database file
and place it elsewhere in the project. I will save any scripts/ files in
the database folder just for organization.