Added quickfix to package.json
Using nw.js we can use the sqlite3 functions in our application. But
because they are BUILT using the quickfix, THEY DO NOT WORK IN A
BROWSER. As such they are not implimented, as it breaks the program for
now.

You can install the program currently by running these two commands back
to back (IN ORDER!)
`npm install`
`npm run bradsQuickFix`

This may change in the future.