Created new database service, replacing old one.
The new service is named the same to angular, but the file is now called
database. I will leave the older databaseInitialization file for
reference. I did a little testing and It seems like the service works as
intended. Need to expand a bit and add error handling, but for now a
user can log in and log out.

NOTE: there is no mention of a userID. I left this out for simplicity,
but now I wonder if we should even take userIDs anymore. If we are to do
so it will be easy to add into the program.