Added jsdoc comments callback check
Now all existing functions have extensive jsdoc comments for all their
arguments. Due to utilizing callbacks for all the functions, there is no
return parameter normally. I also provied a callback check and a return
if there isn't a callback function defined.

I am unsure of how the return will work if you do not define the
callback. It is possible the data will be gone after you leave the
scope, but I did it anyways. If we do find it is useless we can remove
the return statements and just utilize the optional callback functions.

EDIT: Also removed the reserveWords part of isValid name, as the
original code was masking any valid usernames. This function needs to be
changed, and will be at a later time. Until then commented the changes
out

EDIT2: Also added significant database refactoring. This is a rather
large commit, but hopefully one of the last.