Fix for new user creation error
This commit patches the issue where if you where to enter a new user in
the database, the database would fail to add the user and throw and
error. This was due to the fact I was checking if the user is undefined,
not null, which is what the localForage addon returns.