Version 0.1.0
Bare bones essentials are in this release. There is a flawed admin
functionality kept (you don't need a password) and enough UI
responsiveness to work with the program. But a number of features are
missing and or need upgrading. But the main clocking in and clocking are
finished. Going to work on documentation of the working parts and a code
cleanup now. But the first milestone of development has been reached and
finished.