Added momentjs and angular-momentjs support
Adding these libraries will provide us with an easier way of handling
dates and times. The most important aspect of this is the differences in
times saved in the database and the current time.

Next I am going to create the time interfaces.
It is important that you re-download node_modules after this patch