Added new datastructure to handle last logins
A new datastructured called clockedInTimes keeps an array of objects
that consist of a username that corresponds to the username of the user
in that room, and a timeIn, which is used later to display last logged
in time.

This is incorporated into buildOccupied rooms function, which checks the
database integrity of this reserved key. (using the same process as the
clockedInUsers) Also we added this integrity check to ClockIn and
ClockOut functions which keep track of said datastructure. Each function
adds or removes entries.

Also removed uneccessary testDatabase function.