Migrated success clockoutModal to shared
This way I can use the same exact code in two parts of the application.
The only issue is the ID tags must be regulated between the two usages.
Right now there is also a bug that localForage is complaining about when
clocking out. It is unknown if this commit has any effect, positive or
negative for it.