quickfixed swap
Warning this quickfix gets it working, but uses a timeout handle to
prevent a race condition in the database. I am not sure why this
happenes, but it does and thus I utilize a timeout function set to half
a second to wait for the swap to be completed before trying to log in
the other user.