v0.1.19 Test Deployment release
All functionality from the original roomRenter is integrated, and then
some. UI can be upgraded and userIDs need to be checked. As currently
only the last userID entered is kept. This can be changed to be
enforced, or kept track of on a single entry. Also the database printout
can be upgraded.

Also removed the admin password quicksubmit, so now the user must enter
the admin password.