Added new get roomsLoggedIn, also changed getUsers
I changed what the return for getUsers logged in looks like, for easier
parsing. Before it returned an object, now it returns a list of users as
expected.

GetRoomsLoggedin provides an easy way to get all the last logins for
users. Due to the changes on how we save rooms, the name the logged in
with will be automatically provided which is great.